// src/enums/SocketEventNames.ts
export enum SocketEvent {
  GameState = 'game_state',
  GameStateUpdate = 'game_state_update',
  GameEvents = 'game_events',
  Error = 'error',
  JoinRoom = 'join_room',
  ReconnectPlayer = 'reconnect_player',
  StartGame = 'start_game',
  Heartbeat = 'heartbeat',
  RollDice = 'roll_dice',
  ProposeTrade = 'propose_trade',
  AcceptTrade = 'accept_trade',
  CancelTrade = 'cancel_trade',
  DeclineTrade = 'decline_trade',
  BuyProperty = 'buy_property',
  EndTurn = 'end_turn',
  BuildHouse = 'build_house',
  SellHouse = 'sell_house',
  SellProperty = 'sell_property',
  MortgageProperty = 'mortgage_property',
  UnmortgageProperty = 'unmortgage_property',
  PayBail = 'pay_bail',
  UseGooJfCard = 'use_goojf_card',
  LeaveGame = 'leave_game',
  StartAuction = 'start_auction',
  PlaceBid = 'place_bid',
  PassBid = 'pass_bid',
  AddBot = 'add_bot',
  Disconnect = 'disconnect',
  Event = 'event',
  CREATE_GAME = "CREATE_GAME",
  SendMessage = 'send_message',
  NewMessage = 'new_message',
  // Add any other event names used in socket.on or io.emit
}